feat: add StructuredQuery to styles and storybook
