+++ packages/discovery-styles/scss/index.scss
@@ -26,3 +26,4 @@ $css--use-layer: true !default;
 @import 'components/ci-document/ci-document-toolbar';
 @import 'components/ci-document/ci-document-virtual-scroll';
 @import 'components/ci-document/ci-document';
+@import 'components/structured-query/structured-query';
